+++ packages/sdk/test/evm/gasless.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 import { ThirdwebSDK } from "../../src/evm";
-import { assert, expect } from "chai";
+import { assert } from "chai";
 import { ethers, Wallet } from "ethers";
 
 const RPC_URL = "https://rpc-mumbai.maticvigil.com/";
@@ -25,7 +25,6 @@ describe("Gasless Forwarder", async () => {
   });
 
   it.skip("should use sdk with openzeppelin defender", async () => {
-    const NFT_COLLECTION_ADDRESS = "0x24A5aB2878B63716B001aa9AbE816c2662192B12";
     const provider = ethers.getDefaultProvider(RPC_URL);
     const wallet = Wallet.createRandom().connect(provider);
     const sdk = new ThirdwebSDK(wallet, {